Add Fulltext support to Test::Unit
In the Rspec version, :fulltext and :keywords can be used
interchangeably in the tests. But the Test::Unit version only allowed
:keywords. Trying to use :fulltext resulted in an error.

This commit adds a test for :fulltext in the Test::Unit version,
implements :fulltext support and tweaks the indenting in the Test::Unit
file. The last one is only because Syntastic bugged me about it, though.